requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.23
protobuf==3.15.8
PyMySQL==1.0.2
python-dotenv==0.19.0
requests==2.26.0
six==1.15.0
SQLAlchemy==1.4.25
urllib3==1.26.7
Werkzeug==1.0.1
